LPS-172259 Move to the same package of their superclass
