+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/DDMExpressionImpl.java
@@ -22,7 +22,10 @@ import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunctionFactory;
 import com.liferay.dynamic.data.mapping.expression.DDMExpressionFunctionRegistry;
 import com.liferay.dynamic.data.mapping.expression.DDMExpressionObserver;
 import com.liferay.dynamic.data.mapping.expression.DDMExpressionParameterAccessor;
+import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionDSLExpressionVisitor;
+import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionEvaluatorVisitor;
 import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionLexer;
+import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionModelVisitor;
 import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionParser;
 import com.liferay.dynamic.data.mapping.expression.model.Expression;
