+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/parser/DDMExpressionDSLExpressionVisitor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.expression.internal;
+package com.liferay.dynamic.data.mapping.expression.internal.parser;
 
 import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionBaseVisitor;
 import com.liferay.dynamic.data.mapping.expression.internal.parser.DDMExpressionParser;
@@ -330,7 +330,7 @@ public class DDMExpressionDSLExpressionVisitor
 		return values;
 	}
 
-	protected DDMExpressionDSLExpressionVisitor(Map<String, Object> variables) {
+	public DDMExpressionDSLExpressionVisitor(Map<String, Object> variables) {
 		_variables = variables;
 	}
